Add files via upload
Changed the thread length calculation logic--now it calculates the length of the stitches, then multiplies that by two (since the thread will be held double when actually working) and adds an extra 8 inches to the total to account for tying off/floating between sections. The total is still output as inches, and it is rounded up to an integer using math.ceil(). This change has only been made to the GUI version of the application, as that is what will be submitted in the final project.